+++ src/features/manga/components/MangaToolbarMenu.tsx
@@ -94,7 +94,7 @@ export const MangaToolbarMenu = ({ manga, onRefresh, refreshing }: IProps) => {
                                         manga.id,
                                         manga.title,
                                     )}
-                                    state={{ mangaTitle: manga.title }}
+                                    state={{ title: t`Migrate "${manga.title}"` }}
                                     style={{ textDecoration: 'none', color: 'inherit' }}
                                 >
                                     <IconButton color="inherit">
@@ -168,7 +168,7 @@ export const MangaToolbarMenu = ({ manga, onRefresh, refreshing }: IProps) => {
                                     manga.id,
                                     manga.title,
                                 )}
-                                state={{ mangaTitle: manga.title }}
+                                state={{ title: t`Migrate "${manga.title}"` }}
                                 style={{ textDecoration: 'none', color: 'inherit' }}
                             >
                                 <ListItemIcon>