+++ src/features/migration/MigrationManager.ts
@@ -321,7 +321,7 @@ export class MigrationManager {
             ReactRouter.navigate(
                 AppRoutes.migrate.childRoutes.singleMangaSearch.path(manga.sourceId, manga.id, manga.title),
                 {
-                    state: { mangaTitle: manga.title },
+                    state: { title: t`Migrate "${manga.title}"` },
                 },
             );
 
@@ -628,12 +628,26 @@ export class MigrationManager {
                     MigrationEntryStatus.SEARCH_COMPLETE,
                 ].includes(entry.status);
 
-                MigrationManager.selectMatch(mangaId, match.id, match.sourceId, true);
+                if (isSearching) {
+                    draft.searchProgress.completed += 1;
+                    draft.searchProgress.success += 1;
+                    entry.status = MigrationEntryStatus.SEARCH_COMPLETE;
+                }
+
+                entry.isManualSelection = true;
+                entry.selectedMatchMangaId = match.id;
+                entry.selectedMatchSourceId = match.sourceId;
 
                 if (isSearching) {
                     MigrationManager.searchAbortControllerByManga.get(mangaId)?.abort('Manual match selected');
                 }
             }
+        });
+    }
+
+    static openManualSearch(mangaId: MangaIdInfo['id'], title: string): void {
+        ReactRouter.navigate(AppRoutes.migrate.childRoutes.manualSearch.path(mangaId, title), {
+            state: { title: t`Manual migration search for "${title}"` },
         });
     }