+++ src/features/migration/screens/MigrationSelectDestinationSources.tsx
@@ -53,6 +53,7 @@ export const MigrationSelectDestinationSources = () => {
         const filteredSources = Sources.filter(allSources, {
             languages: browseLanguages,
             isNsfw: showNsfw ? undefined : false,
+            keepLocalSource: true,
         });
 
         return Object.values(Sources.groupByLanguage(filteredSources)).flat();
@@ -165,6 +166,7 @@ export const MigrationSelectDestinationSources = () => {
             <Fab
                 variant="extended"
                 color="primary"
+                disabled={!selectedItemIds.length}
                 sx={{
                     position: 'fixed',
                     bottom: (theme) => theme.spacing(2),