+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 - (**Migration**) Add a search option to ignore outdated matches
 - (**Migration**) Add a search option to ignore matches with missing chapters
+- (**Migration**) Add "local source" as a possible destination source
 
 ### Changed
 
@@ -32,6 +33,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - (**Migration**) Fix showing no info message when no sources are available to select as destinations
 - (**Migration**) Fix duplicated search matches after resuming the migration search
 - (**Migration**) Fix search progress not getting updated after a manual search
+- (**Migration**) Fix going back to the migration page after a manual search with a changed search string
+- (**Migration**) Fix being able to start search without selected destination sources
+- (**Migration**) Fix search hotkey (ctrl+f) in the single manga migration search page not focusing the search textfield
 - (**Reader**) Fix scrollbar appearing with "fit to widt/height/screen" page scale mode and applied safe area insets
 - (**Reader**) Fix wrongly positioned mobile progress bar current page indicator
 - (**Reader**) Fix mobile progress bar previous/next chapter button visibility on hover and while disabled